Added welcome layout side bar
